materials_lesson/06.fastAPI_auth/pyproject.toml
[project]
name = "01-create-endpointfastapi-withalembicsqlmodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "scalar-fastapi>=1.6.1",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
